Refactor Luna class to use Storage for task loading and saving
