+++ src/main/java/Luna.java
@@ -1,10 +1,5 @@
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.Scanner;
+
 /**
  * Application entry point
  */
@@ -14,9 +9,10 @@ public class Luna {
     public static void main(String[] args) {
         ArrayList<Task> tasks = new ArrayList<>();
         Ui ui = new Ui();
+        Storage storage = new Storage(DATA_FILE_PATH);
 
         // Load tasks from file at startup
-        loadTasks(tasks);
+        tasks = storage.load();
 
         ui.showWelcome();
 
@@ -35,7 +31,7 @@ public class Luna {
                     case "mark":
                         if (parts.length > 1 && !parts[1].isBlank()) {
                             markCommand(parts[1], tasks, true, ui);
-                            saveTasks(tasks);
+                            storage.save(tasks);
                         } else {
                             throw new LunaException("Please provide a task number to mark");
                         }
@@ -43,7 +39,7 @@ public class Luna {
                     case "unmark":
                         if (parts.length > 1 && !parts[1].isBlank()) {
                             markCommand(parts[1], tasks, false, ui);
-                            saveTasks(tasks);
+                            storage.save(tasks);
                         } else {
                             throw new LunaException("Please provide a task number to unmark");
                         }
@@ -61,7 +57,7 @@ public class Luna {
                             }
                             Task removed = tasks.remove(index);
                             ui.showTaskDeleted(removed, tasks.size());
-                            saveTasks(tasks);
+                            storage.save(tasks);
                         } else {
                             throw new LunaException("Please give a task number to delete");
                         }
@@ -70,19 +66,19 @@ public class Luna {
                         Task todo = new ToDoTask(parts.length > 1 ? parts[1] : "");
                         tasks.add(todo);
                         ui.showTaskAdded(todo, tasks.size());
-                        saveTasks(tasks);
+                        storage.save(tasks);
                         break;
                     case "deadline":
                         Task deadline = new DeadlineTask(parts.length > 1 ? parts[1] : "");
                         tasks.add(deadline);
                         ui.showTaskAdded(deadline, tasks.size());
-                        saveTasks(tasks);
+                        storage.save(tasks);
                         break;
                     case "event":
                         Task event = new EventTask(parts.length > 1 ? parts[1] : "");
                         tasks.add(event);
                         ui.showTaskAdded(event, tasks.size());
-                        saveTasks(tasks);
+                        storage.save(tasks);
                         break;
                     default:
                         throw new LunaException("Sorry! I dont gets");
@@ -95,110 +91,6 @@ public class Luna {
 
         ui.showGoodbye();
         ui.close();
-    }
-
-    /**
-     * Saves tasks to the data file
-     */
-    private static void saveTasks(ArrayList<Task> tasks) {
-        try {
-            // Create data directory if it doesn't exist
-            File dataDir = new File("./data");
-            if (!dataDir.exists()) {
-                dataDir.mkdirs();
-            }
-            FileWriter writer = new FileWriter(DATA_FILE_PATH);
-            for (Task task : tasks) {
-                writer.write(task.taskView() + "\n");
-            }
-            writer.close();
-        } catch (IOException e) {
-            System.out.println("Error saving tasks: " + e.getMessage());
-        }
-    }
-
-    /**
-     * Loads tasks from the data file
-     */
-    private static void loadTasks(ArrayList<Task> tasks) {
-        try {
-            if (!Files.exists(Paths.get(DATA_FILE_PATH))) {
-                return; // No file to load from
-            }
-
-            Scanner fileScanner = new Scanner(new File(DATA_FILE_PATH));
-            while (fileScanner.hasNextLine()) {
-                String line = fileScanner.nextLine();
-                Task task = parseTaskFromFile(line);
-                if (task != null) {
-                    tasks.add(task);
-                }
-            }
-            fileScanner.close();
-        } catch (IOException e) {
-            System.out.println("Error loading tasks: " + e.getMessage());
-        }
-    }
-
-    /**
-     * Parses a task from taskView format
-     */
-    private static Task parseTaskFromFile(String line) {
-        if (line == null || line.trim().isEmpty()) {
-            return null;
-        }
-
-        try {
-
-            if (!line.startsWith("[") || line.length() < 8) {
-                return null;
-            }
-
-            // Extract the content from "[T] [X] ..."
-            char taskType = line.charAt(1);
-            boolean isDone = line.charAt(5) == 'X';
-            String content = line.substring(8);
-
-            Task task = null;
-            switch (taskType) {
-            case 'T':
-                task = new ToDoTask(content);
-                break;
-            case 'D':
-                // Format: description (by: date)
-                int byIndex = content.lastIndexOf(" (by: ");
-                if (byIndex != -1 && content.endsWith(")")) {
-                    String description = content.substring(0, byIndex);
-                    String date = content.substring(byIndex + 6, content.length() - 1);
-                    task = new DeadlineTask(description + " /by " + date);
-                }
-                break;
-            case 'E':
-                // Format: description (from: start to: end)
-                int fromIndex = content.lastIndexOf(" (from: ");
-                if (fromIndex != -1 && content.endsWith(")")) {
-                    String description = content.substring(0, fromIndex);
-                    String timeInfo = content.substring(fromIndex + 8, content.length() - 1);
-                    int toIndex = timeInfo.lastIndexOf(" to: ");
-                    if (toIndex != -1) {
-                        String startTime = timeInfo.substring(0, toIndex);
-                        String endTime = timeInfo.substring(toIndex + 5);
-                        task = new EventTask(description + " /from " + startTime + " /to " + endTime);
-                    }
-                }
-                break;
-            default:
-                return null;
-            }
-
-            if (task != null && isDone) {
-                task.markDone(true);
-            }
-            return task;
-        } catch (LunaException e) {
-            // Skip invalid tasks
-            return null;
-        }
     }
 
     /**
